requirements.txt
streamlit==1.29.0
python-dotenv==1.0.0
requests==2.31.0
pandas==2.1.4
beautifulsoup4==4.12.2
firebase-admin==6.3.0
Pillow==10.1.0
llama-cpp-python==0.2.23
huggingface-hub==0.19.4
instaloader==4.10.1
screaming-frog-seo-spider==18.6
python-canva==0.1.1
google-auth==2.25.2
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.111.0